parsing the list of tags

+++ src/main/kotlin/fr/jfo/examples/bnp/model/Tag.kt
@@ -1,3 +1,3 @@
 package fr.jfo.examples.bnp.model
 
-data class Tag(val unixTimestamp: Long, val customerId: Long, val station: String)
+data class Tag(val unixTimestamp: Long, val customerId: Int, val station: String)
